+++ src/java/com/threerings/tudey/data/effect/Effect.java
@@ -19,7 +19,7 @@ public abstract class Effect extends SimpleStreamableObject
     /**
      * Creates a new effect.
      */
-    public Effect (long timestamp)
+    public Effect (int timestamp)
     {
         _timestamp = timestamp;
     }
@@ -34,7 +34,7 @@ public abstract class Effect extends SimpleStreamableObject
     /**
      * Returns the time at which the effect was fired.
      */
-    public long getTimestamp ()
+    public int getTimestamp ()
     {
         return _timestamp;
     }
@@ -43,7 +43,7 @@ public abstract class Effect extends SimpleStreamableObject
      * Returns the time at which this effect expires (that is, when it should no longer be
      * retransmitted to clients).
      */
-    public long getExpiry ()
+    public int getExpiry ()
     {
         return _timestamp + getLifespan();
     }
@@ -66,11 +66,11 @@ public abstract class Effect extends SimpleStreamableObject
      * Returns the lifespan of the effect, which is the amount of time to continue retransmitting
      * the effect to clients before discarding it.
      */
-    protected long getLifespan ()
+    protected int getLifespan ()
     {
-        return 3000L;
+        return 3000;
     }
 
     /** The time at which the effect was fired. */
-    protected long _timestamp;
+    protected int _timestamp;
 }
