+++ src/java/com/threerings/tudey/data/effect/TransientEffect.java
@@ -20,7 +20,7 @@ public abstract class TransientEffect extends Effect
     /**
      * Creates a new transient effect.
      */
-    public TransientEffect (long timestamp)
+    public TransientEffect (int timestamp)
     {
         super(timestamp);
     }
